SCRUM-4853: Swap AlleleVariantSequence to SequenceSummaryDocument
Replace the old Neo4j-based AlleleVariantSequence with curation API
SequenceSummaryDocument for the gene allele-variant-detail endpoint.
Update API return types, JsonView annotations, ES query to use
sequence_summary category, and fix SIFT/PolyPhen VCF parsing to read
both prediction and score columns separately.

+++ agr_api/src/main/java/org/alliancegenome/api/controller/DiseaseController.java
@@ -609,7 +609,7 @@ public class DiseaseController implements DiseaseRESTInterface {
 	}
 
 	@Override
-	public Integer getCountsOfDiseaseAnnotationsByAllele(String diseaseID) {
+	public Long getCountsOfDiseaseAnnotationsByAllele(String diseaseID) {
 		String associationType = diseaseESService.getAT("allele_disease_annotation", diseaseID);
 		String sortBy = "diseaseAlleleDefault";
 		
@@ -618,7 +618,7 @@ public class DiseaseController implements DiseaseRESTInterface {
 	}
 
 	@Override
-	public Integer getCountsOfDiseaseAnnotationsForModel(String diseaseID) {
+	public Long getCountsOfDiseaseAnnotationsForModel(String diseaseID) {
 		String associationType = diseaseESService.getAT("agm_disease_annotation", diseaseID);
 		
 		JsonResultResponse<AGMDiseaseAnnotationDocument> response = getDiseaseAnnotationsForModel(diseaseID, null, null, null, null, null, null, null, null, null, null, associationType, null, null, null, null, null);
@@ -626,7 +626,7 @@ public class DiseaseController implements DiseaseRESTInterface {
 	}
 
 	@Override
-	public Integer getCountsOfDiseaseAnnotationsByGene(String diseaseID) {
+	public Long getCountsOfDiseaseAnnotationsByGene(String diseaseID) {
 		String associationType = diseaseESService.getAT("gene_disease_annotation", diseaseID);
 		
 		JsonResultResponse<GeneDiseaseAnnotationDocument> response = getDiseaseAnnotationsByGene(diseaseID, null, null, null, null, null, null, null, null, null, null, null, associationType, null, null);
